feat(record): support multiple sports
